add: level chunk management
That is, automatic division and management of a level in chunks
for the purpose of being able to handle many more objects at once.

Before this, whenever the level moved (i. e., scrolled), all objects
in the level were moved and then checked for collision with the screen
to determine which should be drawn.

Now the level is divided in chunks. Only the chunks are moved and then
they are checked for collision with the screen (actually, we use an
area that expands beyond the screen). Only such chunks have their
objects checked for collision with the screen, making everything much
more efficient than checking all objects in the level.

Implementing chunk creation and management didn't require us to change
the data model either, cause the chunks are only created and used when
the app is live.