fix: bug when drawing level for exporting as PNG
We were drawing objects chunk by chunk then layer by layer from each
chunk when we should be doing the opposite: layer by layer then chunk
by chunk.

That is, we must first draw the first layer from each chunk, then the
next layer from each chunk and so on.

The reason for that is that there may be overlap between objects in
neighbooring chunks, causing objects from one chunk to be drawn over
an object from another chunk even when the first object is from a layer
that should never be drawn over the layer of the second object.

By drawing all objects in a layer from all chunks first, and only then
proceeding to the next layer we prevent the issue.